feat: add ShareButton component
